Renamed CommunitiesPage comp to ExplorePage
