--- packages/commonwealth/client/scripts/views/pages/Communities/index.ts
@@ -1,1 +1,0 @@
-export { default } from './Communities';
