+++ packages/commonwealth/client/scripts/views/pages/ExplorePage/CommunitiesList/FiltersDrawer/FiltersDrawer.scss
