+++ packages/commonwealth/client/scripts/views/pages/ExplorePage/ExplorePage.scss
@@ -4,7 +4,7 @@
   z-index: 0 !important;
 }
 
-.CommunitiesPageLayout {
+.ExplorePageLayout {
   padding-top: 0;
   height: 100%;
   overflow-y: auto;
@@ -15,7 +15,7 @@
   }
 }
 
-.CommunitiesPage {
+.ExplorePage {
   display: flex;
   flex: 1;
   flex-direction: column;
