+++ packages/commonwealth/client/scripts/views/pages/ExplorePage/IdeaLaunchpad/TokenLaunchDrawer/index.tsx
