+++ packages/commonwealth/client/scripts/views/pages/ExplorePage/QuestList/QuestCard/QuestCard.scss
