+++ packages/commonwealth/client/scripts/views/pages/ExplorePage/index.ts
@@ -1,0 +1,1 @@
+export { default } from './ExplorePage';
